+++ server/router/battle.js
@@ -1,4 +1,4 @@
-const crud_battle = require('../crud/battle');
+const crud_attack = require('../crud/interact/attack');
 const crud_spawn = require('../crud/user/spawn');
 const crud_interact = require('../crud/interact/interact');
 const config = require('../config');
@@ -10,13 +10,13 @@ module.exports = {
 function add_routes(socket, io) {
     socket.on('suicide', function(data) {
         socket.send({data: "You have died."});
-        crud_battle.create_corpse(socket);
+        crud_attack.create_corpse(socket);
         crud_interact.announce(socket.id, io, 'committed suicide', config.SEEING_DISTANCE, false);
         crud_spawn.respawn(socket.id, io, 'suicide');
     });
 
     socket.on('punch', function(data) {
-        crud_interact.attack_nearby(
+        crud_attack.attack_nearby(
             socket, io, 
             config.ATTACK_DISTANCE,
             config.PUNCH_ENERGY, config.PUNCH_DAMAGE, true
